+++ src/main/java/org/example/ch3_develop/user/controller/UserController.java
@@ -1,5 +1,7 @@
 package org.example.ch3_develop.user.controller;
 
+import jakarta.servlet.http.HttpSession;
+import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.example.ch3_develop.schedule.dto.*;
 import org.example.ch3_develop.schedule.service.ScheduleService;
@@ -17,12 +19,27 @@ import java.util.List;
 public class UserController {
     private final UserService userService;
 
+    // 회원가입
     @PostMapping
-    public ResponseEntity<CreateUserResponse> createUser(
+    public ResponseEntity<CreateUserResponse> signUp(
             @RequestBody CreateUserRequest request){
-        CreateUserResponse result = userService.createUser(request);
+        CreateUserResponse result = userService.signUp(request);
 
         return ResponseEntity.status(HttpStatus.CREATED).body(result);
+    }
+
+    //로그인 로그아웃
+    @PostMapping("/login")
+    public ResponseEntity<Void> login(@Valid @RequestBody LoginRequest request, HttpSession session){
+        SessionUser sessionUser = userService.login(request);
+        session.setAttribute("LoginUser", sessionUser);
+        return ResponseEntity.status(HttpStatus.OK).build();
+    }
+    @PostMapping("/logout")
+    public ResponseEntity<Void> logout(HttpSession session){
+        session.invalidate();
+
+        return ResponseEntity.status(HttpStatus.OK).build();
     }
 
     @GetMapping
@@ -41,7 +58,11 @@ public class UserController {
     @PutMapping("/{id}")
     public ResponseEntity<UpdateUserResponse> UpdateUser(
             @PathVariable Long id,
-            @RequestBody UpdateUserRequest request){
+            @RequestBody UpdateUserRequest request,
+            HttpSession session){
+        if(session.getAttribute("LoginUser") == null){
+            throw new IllegalArgumentException("로그인이 필요합니다");
+        }
         UpdateUserResponse result = userService.updateUser(id, request);
         return ResponseEntity.status(HttpStatus.OK).body(result);
     }
@@ -49,7 +70,11 @@ public class UserController {
     @DeleteMapping("/{id}")
     public ResponseEntity<Void> DeleteUser(
             @PathVariable Long id,
-            @RequestBody DeleteUserRequest request){
+            @RequestBody DeleteUserRequest request,
+            HttpSession session){
+        if(session.getAttribute("LoginUser") == null){
+            throw new IllegalArgumentException("로그인이 필요합니다");
+        }
         userService.deleteUser(id, request);
         return ResponseEntity.status(HttpStatus.NO_CONTENT).build();
     }
